Link the GLFW Window Event to DH Event

+++ D_Hero_Prj/src/DHero/Events/KeyCodes.h
@@ -2,7 +2,7 @@
 
 namespace DH
 {
-	using KeyCode = uint16_t;
+	using KeyCode = unsigned int;
 
 	namespace Key
 	{
